Mariabackup: Disable --innodb-doublewrite by default
xtrabackup --backup does not make any use of the doublewrite buffer.
(This means that it may be unable to back up a crashed database if the
server is offline.)

xtrabackup --prepare in MariaDB 10.1 did not use the doublewrite
buffer. So, in 10.2 it should not be used by default either,
as it is only causing extra writes and slowing down the restore
operation. This means that if xtrabackup --prepare is killed and
restarted, the files might end up being ruined, except if
--innodb-doublewrite was enabled. Note that the incremental restore
(applying .delta files) is not protected by the doublewrite buffer.